failed if error

+++ src/CallWebhookJob.php
@@ -38,6 +38,8 @@ class CallWebhookJob implements ShouldQueue
 
     public bool $verifySsl;
 
+    public bool $failedIfError;
+
     /** @var string|null */
     public $queue = null;
 
@@ -107,12 +109,17 @@ class CallWebhookJob implements ShouldQueue
             }
 
             $this->dispatchEvent(WebhookCallFailedEvent::class);
-        }
 
-        if ($lastAttempt) {
-            $this->dispatchEvent(FinalWebhookCallFailedEvent::class);
+            if ($lastAttempt) {
 
-            $this->delete();
+                $this->dispatchEvent(FinalWebhookCallFailedEvent::class);
+
+                if ($this->failedIfError) {
+                    $this->fail();
+                } else {
+                    $this->delete();
+                }
+            }
         }
     }
